refactor(popup): move source into utils/ and tests into tests/
