+++ entrypoints/popup/Images.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
-  import type { RawImage, ImageStatus } from './types';
-  import type { AltState } from './images';
-  import { altState, fileLabel, imageStatus } from './images';
-  import { csvField, formatSize } from './format';
+  import type { RawImage, ImageStatus } from './utils/types';
+  import type { AltState } from './utils/images';
+  import { altState, fileLabel, imageStatus } from './utils/images';
+  import { csvField, formatSize } from './utils/format';
   import SummaryBar from './SummaryBar.svelte';
   import type { SummaryItem } from './SummaryBar.svelte';
-  import { highlightImages, scrollToImage } from './utils';
+  import { highlightImages, scrollToImage } from './utils/utils';
   import { trackAction } from '@/utils/analytics';
   import { untrack, onDestroy } from 'svelte';
