+++ entrypoints/popup/utils/assets.ts
