+++ entrypoints/popup/utils/dom-text.ts
