+++ entrypoints/popup/utils/format.ts
