+++ entrypoints/popup/utils/headings.ts
