+++ entrypoints/popup/utils/images.ts
