+++ entrypoints/popup/utils/themeStoreLookup.ts
