+++ entrypoints/popup/utils/types.ts
