+++ entrypoints/popup/utils/utils.ts
